+++ src/index.ts
@@ -20,30 +20,31 @@ import { DiagnosticsManager, OutputChannel } from './scripts/output';
 	// Now provide the implementation of the command with  registerCommand
 	// The commandId parameter must match the command field in package.json
 	// prepare and register commands
-	let init = vscode.commands.registerCommand('aurahelper.init', commands.Initialization.run);
-	let addJSFunction = vscode.commands.registerCommand('aurahelper.completion.js.function', commands.AddJSFunction.run);
-	let addMethodBlock = vscode.commands.registerCommand('aurahelper.completion.documentation.method', commands.AddMethodBlock.run);
-	let editApexCommentTemplate = vscode.commands.registerCommand('aurahelper.template.apex.comment', commands.EditApexCommentTemplate.run);
-	let editAuraDocBaseTemplate = vscode.commands.registerCommand('aurahelper.template.aura.documentation', commands.EditAuraDocumentationTemplate.run);
-	let genAuraDoc = vscode.commands.registerCommand('aurahelper.completion.aura.documentation', commands.GenerateAuraDocumentation.run);
-	let newAuraFile = vscode.commands.registerCommand('aurahelper.file.new.aura', commands.NewAuraFile.run);
-	let refreshMetadataIndex = vscode.commands.registerCommand('aurahelper.metadata.refresh.index', commands.RefreshAllMetadataIndex.run);
-	let refreshMetadataIndexForObject = vscode.commands.registerCommand('aurahelper.metadata.refresh.object', commands.RefreshObjectMetadataIndex.run);
-	let compressXML = vscode.commands.registerCommand('aurahelper.metadata.compress.xml', commands.CompressXML.run);
-	let compressAllXML = vscode.commands.registerCommand('aurahelper.metadata.compress.xml.all', commands.CompressAllXML.run);
-	let permissionEditor = vscode.commands.registerCommand('aurahelper.metadata.permission.editor', commands.OpenPermissionEditor.run);
-	let retrieveSpecialTypes = vscode.commands.registerCommand('aurahelper.metadata.retrieve.special', commands.RetrieveSpecialTypes.run);
-	let packageGenerator = vscode.commands.registerCommand('aurahelper.metadata.package.generate', commands.OpenPackageGeneratorGUI.run);
-	let orgCompare = vscode.commands.registerCommand('aurahelper.metadata.org.comparator', commands.OrgCompare.run);
-	let openCustomLabelsEditor = vscode.commands.registerCommand('aurahelper.metadata.customlabels.editor', commands.OpenCustomLabelsEditor.run);
-	let createProjectDocumentation = vscode.commands.registerCommand('aurahelper.documentation.project.create', commands.CreateProjectDocumentation.run);
-	let implementInterfaces = vscode.commands.registerCommand('aurahelper.completion.apex.implement.interface', commands.ImplementInterfaces.run);
-	let implementExtendedClasses = vscode.commands.registerCommand('aurahelper.completion.apex.implement.extend', commands.ImplementExtendedClasses.run);
-	let repairProjectDependencies = vscode.commands.registerCommand('aurahelper.metadata.project.repair', commands.RepairProjectDependencies.run);
-	let ignoreMetadata = vscode.commands.registerCommand('aurahelper.metadata.ignore', commands.IgnoreMetadata.run);
-	let modifyPermissions = vscode.commands.registerCommand('aurahelper.metadata.permission.modify', commands.ModifyPermissions.run);
-	let help = vscode.commands.registerCommand('aurahelper.help', commands.Help.run);
-	let clearProblemsOutput = vscode.commands.registerCommand('aurahelper.output.problems.clear', function () {
+	const init = vscode.commands.registerCommand('aurahelper.init', commands.Initialization.run);
+	const addJSFunction = vscode.commands.registerCommand('aurahelper.completion.js.function', commands.AddJSFunction.run);
+	const addMethodBlock = vscode.commands.registerCommand('aurahelper.completion.documentation.method', commands.AddMethodBlock.run);
+	const editApexCommentTemplate = vscode.commands.registerCommand('aurahelper.template.apex.comment', commands.EditApexCommentTemplate.run);
+	const editAuraDocBaseTemplate = vscode.commands.registerCommand('aurahelper.template.aura.documentation', commands.EditAuraDocumentationTemplate.run);
+	const genAuraDoc = vscode.commands.registerCommand('aurahelper.completion.aura.documentation', commands.GenerateAuraDocumentation.run);
+	const newAuraFile = vscode.commands.registerCommand('aurahelper.file.new.aura', commands.NewAuraFile.run);
+	const refreshMetadataIndex = vscode.commands.registerCommand('aurahelper.metadata.refresh.index', commands.RefreshAllMetadataIndex.run);
+	const refreshMetadataIndexForObject = vscode.commands.registerCommand('aurahelper.metadata.refresh.object', commands.RefreshObjectMetadataIndex.run);
+	const compressXML = vscode.commands.registerCommand('aurahelper.metadata.compress.xml', commands.CompressXML.run);
+	const compressAllXML = vscode.commands.registerCommand('aurahelper.metadata.compress.xml.all', commands.CompressAllXML.run);
+	const permissionEditor = vscode.commands.registerCommand('aurahelper.metadata.permission.editor', commands.OpenPermissionEditor.run);
+	const retrieveSpecialTypes = vscode.commands.registerCommand('aurahelper.metadata.retrieve.special', commands.RetrieveSpecialTypes.run);
+	const packageGenerator = vscode.commands.registerCommand('aurahelper.metadata.package.generate', commands.OpenPackageGeneratorGUI.run);
+	const orgCompare = vscode.commands.registerCommand('aurahelper.metadata.org.comparator', commands.OrgCompare.run);
+	const openCustomLabelsEditor = vscode.commands.registerCommand('aurahelper.metadata.customlabels.editor', commands.OpenCustomLabelsEditor.run);
+	const createProjectDocumentation = vscode.commands.registerCommand('aurahelper.documentation.project.create', commands.CreateProjectDocumentation.run);
+	const implementInterfaces = vscode.commands.registerCommand('aurahelper.completion.apex.implement.interface', commands.ImplementInterfaces.run);
+	const implementExtendedClasses = vscode.commands.registerCommand('aurahelper.completion.apex.implement.extend', commands.ImplementExtendedClasses.run);
+	const repairProjectDependencies = vscode.commands.registerCommand('aurahelper.metadata.project.repair', commands.RepairProjectDependencies.run);
+	const ignoreMetadata = vscode.commands.registerCommand('aurahelper.metadata.ignore', commands.IgnoreMetadata.run);
+	const modifyPermissions = vscode.commands.registerCommand('aurahelper.metadata.permission.modify', commands.ModifyPermissions.run);
+	const scan = vscode.commands.registerCommand('aurahelper.metadata.scan', commands.Scan.run);
+	const help = vscode.commands.registerCommand('aurahelper.help', commands.Help.run);
+	const clearProblemsOutput = vscode.commands.registerCommand('aurahelper.output.problems.clear', function () {
 		DiagnosticsManager.clearDiagnostics();
 	});
 
@@ -72,6 +73,7 @@ import { DiagnosticsManager, OutputChannel } from './scripts/output';
 	context.subscriptions.push(ignoreMetadata);
 	context.subscriptions.push(modifyPermissions);
 	context.subscriptions.push(init);
+	context.subscriptions.push(scan);
 	context.subscriptions.push(help);
 	console.log('Aura Helper Extension activated');
 	OutputChannel.createChannel();